Finished App No GUI (1/3/24)
This is the finish application, coded in Java, which allows the user to input either a percentage for their grades or a letter grade, which also allows for an early bird class to be added to the list. This should perform without error, and will be imported in phases to an application with a GUI in Android Studio.